Moved check from query to separate command
