+++ tests/cli/test_cli_query.py
@@ -7,8 +7,6 @@ from click.testing import CliRunner
 
 from bigquery_etl.cli.query import (
     _attach_metadata,
-    _parse_check_output,
-    _parse_partition_setting,
     create,
     info,
     paths_matching_name_pattern,
@@ -418,23 +416,3 @@ class TestQuery:
             assert "foo" in table.labels
             assert table.labels["foo"] == "abc"
             assert "review_bugs" not in table.labels
-
-    def test_parse_check_output(self):
-        expected = "ETL Data Check Failed: a check failed"
-        assert _parse_check_output(expected) == expected
-
-        test2 = "remove prepended text ETL Data Check Failed: a check failed"
-        assert _parse_check_output(test2) == expected
-
-        test3 = "no match for text Data Check Failed: a check failed"
-        assert _parse_check_output(test3) == test3
-
-    def test_parse_partition_setting(self):
-        assert _parse_partition_setting("submission_date::2023-06-01") == {
-            "submission_date": "2023-06-01"
-        }
-        assert _parse_partition_setting("submission_date2023-06-01") is None
-        assert _parse_partition_setting("submission_date:2023-06-01") is None
-        assert _parse_partition_setting("submission_date$::2023-06-01") is None
-        assert _parse_partition_setting("submission_date::2023-13-01") is None
-        assert _parse_partition_setting("submission_date::20230601") is None
